# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SP 1.0",
        "description": "SP API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/thematic_areas": {
            "get": {
                "tags": [
                    "thematic_areas"
                ],
                "summary": "Lista as áreas temáticas",
                "description": "Lista todas as áreas temáticas cadastradas no sistema",
                "operationId": "0ce92383f7c24efa912fe14ec289fb44",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Quantidade de itens por página",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordem de exibição dos itens [asc, desc]",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "column",
                        "in": "query",
                        "description": "Coluna a ser utilizada para realizar a ordenação",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar por valor",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Usuário não autenticado"
                    },
                    "403": {
                        "description": "Usuário não tem permissão de acesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "thematic_areas"
                ],
                "summary": "Cadastra áreas temáticas",
                "description": "Cadastra novas áreas temáticas na plataforma",
                "operationId": "b4dce9c9a9db2eec9e3bab1988b1f1ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description"
                                ],
                                "properties": {
                                    "description": {
                                        "description": "Descrição",
                                        "type": "string",
                                        "example": "Test"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Usuário não autenticado"
                    },
                    "403": {
                        "description": "Usuário não tem permissão de acesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/thematic_areas/{id}": {
            "get": {
                "tags": [
                    "thematic_areas"
                ],
                "summary": "Consulta os dados da área temática cadastrada",
                "description": "Consulta com mais detalhes os dados da área temática",
                "operationId": "2117b60c4898c4393f2d99cbde0901ef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da área temática cadastrada",
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Usuário não autenticado"
                    },
                    "403": {
                        "description": "Usuário não tem permissão de acesso"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "thematic_areas"
                ],
                "summary": "Atualiza os dados da área temática",
                "description": "",
                "operationId": "655a45e4f8ba62fde93c73f1a9bf6ea5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da área temática cadastrada",
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description"
                                ],
                                "properties": {
                                    "description": {
                                        "description": "Descrição",
                                        "type": "string",
                                        "example": "Test"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Usuário não autenticado"
                    },
                    "403": {
                        "description": "Usuário não tem permissão de acesso"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "thematic_areas"
                ],
                "summary": "Remove áreas temáticas cadastradas",
                "description": "Remove informações de áreas temáticas cadastradas",
                "operationId": "f38e34fa4110787032144921de1a0fd5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da área temática cadastrada",
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Usuário não autenticado"
                    },
                    "403": {
                        "description": "Usuário não tem permissão de acesso"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "description": "Insira o Personal Access Token assim: Bearer <token>",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ]
}